move list item test and fix it
